+++ engine/tests/mock_simulation.rs
@@ -1,81 +1,18 @@
-use super::*;
-
 const AGENT_ID: &str = "agent";
 
-use std::{pin::Pin, time::Duration};
+use arbiter_macros::Behaviors;
+use futures_util::StreamExt;
 
-use anyhow::Result;
-use ethers::types::BigEndianHash;
-use futures_util::Stream;
-use serde::*;
+use arbiter_engine::{
+    agent::Agent,
+    machine::{CreateStateMachine, Engine, StateMachine},
+    world::World,
+};
+
+use std::time::Duration;
+
 use tokio::time::timeout;
-
-use super::*;
-
-fn default_max_count() -> Option<u64> {
-    Some(3)
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub(crate) struct TimedMessage {
-    delay: u64,
-    receive_data: String,
-    send_data: String,
-    #[serde(skip)]
-    messager: Option<Messager>,
-    #[serde(default)]
-    count: u64,
-    #[serde(default = "default_max_count")]
-    max_count: Option<u64>,
-    startup_message: Option<String>,
-}
-
-impl TimedMessage {
-    pub fn new(
-        delay: u64,
-        receive_data: String,
-        send_data: String,
-        max_count: Option<u64>,
-        startup_message: Option<String>,
-    ) -> Self {
-        Self {
-            delay,
-            receive_data,
-            send_data,
-            messager: None,
-            count: 0,
-            max_count,
-            startup_message,
-        }
-    }
-}
-
-#[async_trait::async_trait]
-impl Behavior<Message> for TimedMessage {
-    async fn startup(
-        &mut self,
-        _client: Arc<ArbiterMiddleware>,
-        messager: Messager,
-    ) -> Result<EventStream<Message>> {
-        if let Some(startup_message) = &self.startup_message {
-            messager.send(To::All, startup_message).await;
-        }
-        self.messager = Some(messager.clone());
-        Ok(messager.stream()?)
-    }
-
-    async fn process(&mut self, event: Message) -> Result<ControlFlow> {
-        if event.data == serde_json::to_string(&self.receive_data).unwrap() {
-            let messager = self.messager.clone().unwrap();
-            messager.send(To::All, self.send_data.clone()).await;
-            self.count += 1;
-        }
-        if self.count == self.max_count.unwrap_or(u64::MAX) {
-            return Ok(ControlFlow::Halt);
-        }
-        Ok(ControlFlow::Continue)
-    }
-}
+include!("common.rs");
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
 async fn echoer() {
@@ -92,7 +29,7 @@ async fn echoer() {
     world.add_agent(agent.with_behavior(behavior));
     let messager = world.messager.for_agent("outside_world");
 
-    world.run().await;
+    world.run().await.unwrap();
 
     let mut stream = messager.stream().unwrap();
     let mut idx = 0;
@@ -134,7 +71,7 @@ async fn ping_pong() {
     );
 
     let messager = world.messager.for_agent("outside_world");
-    world.run().await;
+    world.run().await.unwrap();
 
     let mut stream = messager.stream().unwrap();
     let mut idx = 0;
@@ -175,7 +112,7 @@ async fn ping_pong_two_agent() {
     world.add_agent(agent_pong.with_behavior(behavior_pong));
 
     let messager = world.messager.for_agent("outside_world");
-    world.run().await;
+    world.run().await.unwrap();
 
     let mut stream = messager.stream().unwrap();
     let mut idx = 0;
@@ -203,8 +140,7 @@ enum Behaviors {
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
 async fn config_test() {
-    let mut world =
-        World::from_config::<Behaviors>("src/examples/timed_message/config.toml").unwrap();
+    let mut world = World::from_config::<Behaviors>("tests/config.toml").unwrap();
     assert_eq!(world.id, "timed_message_world");
-    world.run().await;
+    world.run().await.unwrap();
 }
