+++ examples/minter/behaviors/mod.rs
