+++ examples/minter/behaviors/token_admin.rs
