+++ examples/minter/behaviors/token_requester.rs
