+++ examples/minter/mod.rs
