+++ examples/minter/token_minter.rs
